requirements.txt
Django==2.2.6
mistune
mysqlclient
django-debug-toolbar
